extract bullet metric grid to own file
